* CC: fix the cctest regression on cc_function_ptr.cpp.
This is because the Tokenizer::ReadParentheses() function add some
extra spaces in the result, this make the handle function pointer
related code failed.
For example, it return the string
L"(* FuncArray [ 10 ] [ 20 ])". This commit change the string to
old format, which is L"(*FuncArray[10][20])".
See details in: http://forums.codeblocks.org/index.php/topic,19769.msg138614.html#msg138614